requirements.txt
# NeuroQuant Requirements v3.0 - Institutional Backtesting Platform
# No RL/AI dependencies - Pure quantitative finance

# Core data science
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# Financial data and technical analysis
yfinance>=0.2.28
TA-Lib>=0.4.25

# Web framework and API
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6
websockets>=11.0.3
aiohttp>=3.8.5

# Database and caching (optional for future enhancements)
psycopg2-binary>=2.9.9  # PostgreSQL adapter for production
redis>=4.6.0
aioredis>=2.0.1

# Configuration and environment
python-dotenv>=1.0.0

# Utilities
requests>=2.31.0
scikit-learn>=1.3.0
pytz>=2023.3
scipy>=1.11.0
matplotlib>=3.7.0

# Logging and monitoring
colorlog>=6.7.0

# Development and testing tools
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.4.0
httpx>=0.24.1

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.1.0

# Note: TA-Lib requires system-level installation
# On Ubuntu/Debian: sudo apt-get install libta-lib-dev
# On macOS: brew install ta-lib
# On Windows: Download from https://www.lfd.uci.edu/~gohlke/pythonlibs/#ta-lib